unwrap errors ahen checking for errorx type
