+++ type.go
@@ -95,14 +95,6 @@ func (t *Type) HasTrait(key Trait) bool {
 	return ok
 }
 
-// IsOfType is a type check for errors.
-// Returns true either if both are of exactly the same type, or if the same is true for one of current type's ancestors.
-// For an error that does not have an errorx type, returns false.
-func IsOfType(err error, t *Type) bool {
-	e := Cast(err)
-	return e != nil && e.IsOfType(t)
-}
-
 // Supertype returns a parent type, if present.
 func (t *Type) Supertype() *Type {
 	return t.parent
